+++ systems/__init__.py
@@ -1,6 +1,7 @@
 # MACRO core systems modules
 from .mobility_system import MotionController
 from .navigation_system import Transformation, Location, Navigation
+from .state import State
 from .sensors import SensorInput
 
 # Note: task_manager.py is currently empty
@@ -11,4 +12,5 @@ __all__ = [
     'Location',
     'Navigation',
     'SensorInput',
+    'State',
 ]